requirements.txt
absl-py==2.1.0
adbutils==2.9.4
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Automat==25.4.16
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boto3==1.37.34
botocore==1.37.34
bs4==0.0.2
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.16.0
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
deepface==0.0.94
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
dj-database-url==3.0.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.7.0
firebase_admin==7.1.0
Flask==2.0.2
flask-cors==6.0.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.9.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
h5py==3.11.0
hpack==4.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaggle==1.6.17
keras==3.11.1
kiwisolver==1.4.5
langdetect==1.0.9
libclang==18.1.1
lxml==5.4.0
lz4==4.4.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.99.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pillow==10.4.0
pipenv==2024.4.1
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycoingecko==3.2.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
queuelib==1.8.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
retina-face==0.0.17
retry2==0.9.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9.1
s3transfer==0.11.4
scikit-learn==1.5.1
scipy==1.14.0
Scrapy==2.12.0
seaborn==0.13.2
selenium==4.35.0
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==70.3.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-intel==2.17.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.19.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tinycss2==1.3.0
tldextract==5.3.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
uiautomator2==3.4.0
ultralytics==8.2.101
ultralytics-thop==2.0.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.5.0
validators==0.34.0
virtualenv==20.29.3
w3lib==2.3.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.0.2
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.1
zope.interface==7.2
